requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
firebase-admin==7.2.0
python-dotenv==1.1.1
